requirements.txt
asgiref==3.10.0
Django==5.2.7
djangorestframework==3.16.1
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
Pillow==11.0.0
promise==2.3
psycopg==3.2.11
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.15.0
tzdata==2025.2